app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false  # Add this line


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
celery = "^5.5.2"
redis = "^6.0.0"
psycopg2-binary = "^2.9.10"
flower = "^2.0.1"
reportlab = "^4.4.0"
gunicorn = "^21.2.0"
google-generativeai = "^0.3.2"  
python-dotenv = "^1.0.0"        
beautifulsoup4 = "^4.13.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
